Split mysql database tests into two modules to respect naming convention
